Move functions into template-ui
